Make get_property_class Arch-specific
